kafka-admin-api-python/requirements.txt
fastapi[all]==0.110.1
kafka-python==2.0.2
python-dotenv
